+++ app/src/main/java/woowacourse/movie/presentation/extensions/ContextExt.kt
@@ -1,11 +1,9 @@
 package woowacourse.movie.presentation.extensions
 
 import android.content.Context
-import android.content.DialogInterface
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
-import woowacourse.movie.R
 
 fun Context.checkPermissions(vararg permissions: String): Boolean =
     permissions.all { checkPermission(it) }
@@ -26,25 +24,3 @@ fun Context.createAlertDialog(
         setCancelable(isCancelable)
         builder()
     }.create()
-
-fun AlertDialog.Builder.title(text: String) {
-    this.setTitle(text)
-}
-
-fun AlertDialog.Builder.message(text: String) {
-    this.setMessage(text)
-}
-
-fun AlertDialog.Builder.positiveButton(
-    text: String = context.getString(R.string.yes),
-    onClick: (dialogInterface: DialogInterface) -> Unit = {},
-) {
-    this.setPositiveButton(text) { dialogInterface, _ -> onClick(dialogInterface) }
-}
-
-fun AlertDialog.Builder.negativeButton(
-    text: String = context.getString(R.string.no),
-    onClick: (dialogInterface: DialogInterface) -> Unit = {},
-) {
-    this.setNegativeButton(text) { dialogInterface, _ -> onClick(dialogInterface) }
-}
